requirements.txt
alembic==1.1.0
amazon-page-parser==0.1.3
beautifulsoup4==4.8.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
cssselect==1.1.0
curl-http==0.1.2
googletrans==2.4.0
idna==2.8
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
parsel==1.5.2
psycopg2==2.7.5
pycurl==7.43.0.3
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1.2
requests==2.22.0
schedule==0.6.0
six==1.12.0
soupsieve==1.9.3
SQLAlchemy==1.3.8
sqlparse==0.3.0
urllib3==1.25.3
w3lib==1.21.0
